Add delegation to geometry in Feature(Collection)
Both entities Feature and FeatureCollection are now provided a
`method_missing` that will default to inner geometry.

CAVEATS: The inner geometry for a FeatureCollection is a
GeometryCollection which is not exactly true, but a simpler solution
for RGeo compatibility. We can still override methods that behave
poorly directly in `RGeo::GeoJSON::FeatureCollectionœ.